fix(integration): carry SQLite's 8-byte widths into the test harnesses
The four statically typed SQLite harnesses assert against the generated
types, so widening INTEGER to int64 broke their compilation. The Java and
Kotlin harnesses held the new user's id in an int, and Kotlin read
last_insert_rowid() with getInt even though a rowid is 64-bit.

The Rust harness was already broken before this change: 93b76b9 made
SQLite REAL an 8-byte float in 0.12.0, but its template still declared
`let total: f32`, so `order.total == total` could not compile. CI runs
this suite, so it has been red since that commit -- the Actions outage
masked it on the last few runs. Kotlin got the equivalent fix at the
time; Rust and Java did not.

+++ integration_tests/java-jdbc-sqlite/src/main/java/IntegrationTest.java
@@ -74,7 +74,8 @@ public class IntegrationTest {
         }
     }
 
-    private static int createdUserId;
+    // SQLite INTEGER is 8-byte, so generated id columns are long.
+    private static long createdUserId;
 
     private static void testCreateUser(Connection conn) {
         String name = "CreateUser";
